Add rough red and ints.
I've added integers to the AST, but not to values. They're only used
in constructing relations and not for any actual computation; integers
outside of relations (and as arguments to projections and iotas)
remain illegal and won't compile.

Reals now also need to be specified with a period, i.e. `5` is parsed
as an int and `5.0` as a real. We could infer the right type
to avoid this.